requirements.txt
ccxt>=4.3.0
pandas>=2.1.0
numpy>=1.26.0
pandas_ta>=0.3.14b
vectorbt>=0.26.0
optuna>=3.6.0
pyarrow>=15.0.0
plotly>=5.22.0
python-dotenv>=1.0.0
numba>=0.59.0
scipy>=1.13.0
rich>=13.7.0
click>=8.1.0
pyyaml>=6.0.0
kaleido>=0.2.1
